CDCM6208: make ReadRegister failures detectable
The second return in ReadRegister was unreachable, so SPI read errors
were never reported and the callers' == -1 checks could never fire.
Return the value through a signed 32-bit type so -1 is a real sentinel.

Fixes #276

+++ src/chips/CDCM6208/CDCM6208.h
@@ -144,7 +144,7 @@ class LIME_API CDCM_Dev
 
   private:
     int WriteRegister(uint16_t addr, uint16_t val);
-    uint16_t ReadRegister(uint16_t addr);
+    int32_t ReadRegister(uint16_t addr);
 
     int SolveN(int Target, int* Mult8bit, int* Mult10bit);
     void CalculateFracDiv(CDCM_Output* Output);
